[sl-core] Optimize `memcpy', `memmove' and `bcopy'.
* lib/src/bcopy.c: Use `long' instead of `int' as the copy unit.

+++ slc/lib/src/bcopy.c
@@ -45,7 +45,7 @@ __FBSDID("$FreeBSD: src/lib/libc/string/bcopy.c,v 1.7.10.1 2009/08/03 08:13:06 k
  * sizeof(word) MUST BE A POWER OF TWO
  * SO THAT wmask BELOW IS ALL ONES
  */
-typedef	int word;		/* "word" used for optimal copy speed */
+typedef	long word;		/* "word" used for optimal copy speed */
 
 #define	wsize	sizeof(word)
 #define	wmask	(wsize - 1)
